requirements.txt
asgiref==3.11.1
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
pdfplumber==0.10.3
python-docx==1.1.0
pytest==8.0.0
sentence-transformers==3.0.0
torch==2.2.0
psycopg2-binary==2.9.9
numpy==1.26.0
tqdm==4.66.0
sentence-transformers==3.0.0
torch==2.2.0
psycopg2-binary==2.9.9
numpy==1.26.0
tqdm==4.66.0
requests>=2.28.0
psycopg2-binary>=2.9.0
dataclasses>=0.6; python_version<"3.7"
typing-extensions>=4.0.0; python_version<"3.8"
pydantic>=2.5.0
